Poseido cipher - Error implementation
