+++ Cargo.toml
@@ -12,6 +12,8 @@ nstack = "0.4"
 lazy_static = "1.3.0"
 hades252 = { git = "https://github.com/dusk-network/hades252", version = "0.6.1" }
 dusk-plonk = "0.2.6"
+anyhow = "1.0"
+thiserror = "1.0"
 
 [dev-dependencies]
 rand = "0.7"
